# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/application/v1/create": {
            "post": {
                "description": "Create a specific application",
                "tags": [
                    "applications"
                ],
                "summary": "Create Application",
                "parameters": [
                    {
                        "description": "Application body",
                        "name": "application",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/application.CreateApplicationInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/domain.BadResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/domain.BadResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/domain.BadResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/domain.BadResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/domain.BadResponse"
                        }
                    }
                }
            }
        },
        "/api/application/v1/download-file/{:fileName}": {
            "get": {
                "description": "download file",
                "tags": [
                    "applications"
                ],
                "summary": "Download file",
                "parameters": [
                    {
                        "type": "string",
                        "description": "File name",
                        "name": "filename",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/domain.BadResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/domain.BadResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/domain.BadResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/domain.BadResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/domain.BadResponse"
                        }
                    }
                }
            }
        },
        "/api/application/v1/list": {
            "get": {
                "description": "get all applications",
                "tags": [
                    "applications"
                ],
                "summary": "receiving applications",
                "parameters": [
                    {
                        "type": "string",
                        "description": "9999990000",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "enum": [
                                "canceled",
                                "waiting",
                                "working",
                                "done"
                            ],
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "description": "string enums",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "enum": [
                                "general",
                                "financial"
                            ],
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "description": "string enums",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "enum": [
                                "information",
                                "account_help",
                                "refunds",
                                "payment"
                            ],
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "description": "string enums",
                        "name": "subType",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "2019-01-25T10:30:00.000Z",
                        "name": "createdAtFrom",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "2019-02-25T10:30:00.000Z",
                        "name": "createdAtTo",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "2019-01-25T10:30:00.000Z",
                        "name": "updatedAtFrom",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "2019-02-25T10:30:00.000Z",
                        "name": "updatedAtTo",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.ResponseApplications"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/domain.BadResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/domain.BadResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/domain.BadResponse"
                        }
                    },
                    "405": {
                        "description": "Method Not Allowed",
                        "schema": {
                            "$ref": "#/definitions/domain.BadResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/domain.BadResponse"
                        }
                    }
                }
            }
        },
        "/api/application/v1/my-list": {
            "get": {
                "description": "get all applications for user id",
                "tags": [
                    "applications"
                ],
                "summary": "receiving applications",
                "parameters": [
                    {
                        "type": "string",
                        "description": "9999990000",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "enum": [
                                "canceled",
                                "waiting",
                                "working",
                                "done"
                            ],
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "description": "string enums",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "enum": [
                                "general",
                                "financial"
                            ],
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "description": "string enums",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "enum": [
                                "information",
                                "account_help",
                                "refunds",
                                "payment"
                            ],
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "description": "string enums",
                        "name": "subType",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "2019-01-25T10:30:00.000Z",
                        "name": "createdAtFrom",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "2019-02-25T10:30:00.000Z",
                        "name": "createdAtTo",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "2019-01-25T10:30:00.000Z",
                        "name": "updatedAtFrom",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "2019-02-25T10:30:00.000Z",
                        "name": "updatedAtTo",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.ResponseApplications"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/domain.BadResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/domain.BadResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/domain.BadResponse"
                        }
                    },
                    "405": {
                        "description": "Method Not Allowed",
                        "schema": {
                            "$ref": "#/definitions/domain.BadResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/domain.BadResponse"
                        }
                    }
                }
            }
        },
        "/api/application/v1/{id}": {
            "get": {
                "description": "Get a specific application by ID",
                "tags": [
                    "applications"
                ],
                "summary": "Get Application by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Application ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.ResponseApplication"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/domain.BadResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/domain.BadResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/domain.BadResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/domain.BadResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/domain.BadResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "application.CreateApplicationInput": {
            "type": "object"
        },
...
        "domain.Application": {
            "type": "object",
            "properties": {
                "_id": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "fileName": {
                    "type": "string"
                },
                "managerId": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/domain.ApplicationStatus"
                },
                "subType": {
                    "$ref": "#/definitions/domain.ApplicationSubType"
                },
                "ticketResponseId": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "type": {
                    "$ref": "#/definitions/domain.ApplicationType"
                },
                "updatedAt": {
                    "type": "string"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
        "domain.ApplicationStatus": {
            "type": "string",
            "enum": [
                "canceled",
                "waiting",
                "working",
                "done"
            ],
            "x-enum-varnames": [
                "ApplicationStatusCanceled",
                "ApplicationStatusWaiting",
                "ApplicationStatusWorking",
                "ApplicationStatusDone"
            ]
        },
        "domain.ApplicationSubType": {
            "type": "string",
            "enum": [
                "information",
                "account_help",
                "refunds",
                "payment"
            ],
            "x-enum-varnames": [
                "ApplicationSubTypeInformation",
                "ApplicationSubTypeAccountHelp",
                "ApplicationSubTypeRefunds",
                "ApplicationSubTypePayment"
            ]
        },
        "domain.ApplicationType": {
            "type": "string",
            "enum": [
                "general",
                "financial"
            ],
            "x-enum-varnames": [
                "ApplicationTypeGeneral",
                "ApplicationTypeFinancial"
            ]
        },
        "domain.BadResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "domain.Response": {
            "type": "object",
            "properties": {
                "payload": {},
                "status": {
                    "type": "integer"
                }
            }
        },
        "domain.ResponseApplication": {
            "type": "object",
            "properties": {
                "payload": {
                    "$ref": "#/definitions/domain.Application"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "domain.ResponseApplications": {
            "type": "object",
            "properties": {
                "payload": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.Application"
                    }
                },
                "status": {
                    "type": "integer"
                }
            }
        },
...
    }
}